Safari-611.1.10/LayoutTests/media/remote-control-command-is-user-gesture-expected.txt

Test that a remote control command is treated as a user gesture.


RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForVideoRateChange"))
* set video.src
EVENT(loadedmetadata)

* video.play() should fail
RUN(handlePromise(video.play()))
EXPECTED (video.paused == 'true') OK

* Send a play command, it should succeed.
RUN(internals.postRemoteControlCommand('play'))

EVENT(timeupdate)
END OF TEST
